+++ packages/mds-config-service/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2021 City of Los Angeles
+ * Copyright 2022 City of Los Angeles
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-import webpack from '@mds-core/mds-webpack-config'
-
-export default webpack
-  .Bundle('../../packages/mds-ingest-migration-processor', { name: 'processor' })
-  .UsingDefaultConfig()
+export * from './@types'
+export * from './client'
+export * from './service/manager'