+++ container-images/mds-compliance-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-compliance-service
 
+## 0.0.61
+
+### Patch Changes
+
+- @mds-core/mds-compliance-service@0.10.21
+
 ## 0.0.60
 
 ### Patch Changes